refactor(icons): new LuccaIcon type
This file is generated from the selection.json file that icomoon gives us, then exported to be used in the ng package or in other consumers.

BREAKING CHANGE
This removes the icons-list.js file, which, according to a github search, idn't used anywhere, but better safe than sorry so a check might be needed.