[textFlow] Remove margins on both ends of `.textFlow` (#3481)
* [textFlow] Remove margins on both ends of `.textFlow`

To avoid extra margins, we can remove first-child’s start margin and last-child’s end margin.

This is especially useful in isolated context (eg: modals’ content), which seems to be the use-case for `.textFlow`.

Their container can be added extra margins if need be.

* [textFlow] Don’t mix tabs and spaces.

* [textFlow] Use `0` instead of `unset`:

This is closer to the code habits LF is using.

* [textFlow] Remove bottom margin on component…

Addresses https://github.com/LuccaSA/lucca-front/pull/3481#issuecomment-2709977193.